created tests and test environment

+++ database/migrations/2025_01_20_085133_create_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->float('price')->default(0.0);
-            $table->float('quantity')->default(0.0);
+            $table->unsignedInteger('quantity')->default(0);
             $table->timestamps();
         });
     }
